Try to fix new tests
