+++ packages/scipy/test_scipy.py
@@ -1,15 +1,12 @@
-from textwrap import dedent
-
 import pytest
 from conftest import selenium_context_manager
 
 
 @pytest.mark.driver_timeout(40)
-def test_scipy_linalg(selenium_module_scope, request):
+def test_scipy_linalg(selenium_module_scope):
     with selenium_context_manager(selenium_module_scope) as selenium:
-
         selenium.load_package("scipy")
-        cmd = dedent(
+        selenium.run(
             r"""
             import numpy as np
             import scipy as sp
@@ -28,8 +25,6 @@ def test_scipy_linalg(selenium_module_scope, request):
             """
         )
 
-        selenium.run(cmd)
-
 
 @pytest.mark.driver_timeout(40)
 def test_brentq(selenium_module_scope):
@@ -42,24 +37,14 @@ def test_brentq(selenium_module_scope):
             """
         )
 
-@pytest.mark.driver_timeout(40)
-def test_dlamch(selenium_module_scope):
-    with selenium_context_manager(selenium_module_scope) as selenium:
-        from scipy.linalg import lapack
-        lapack.dlamch('Epsilon-Machine')
 
 @pytest.mark.driver_timeout(40)
-def test_logistic_regression(selenium_module_scope):
+def test_dlamch(selenium_module_scope):
     with selenium_context_manager(selenium_module_scope) as selenium:
         selenium.load_package("scipy")
         selenium.run(
             """
-            from sklearn.datasets import load_iris
-            from sklearn.linear_model import LogisticRegression
-            X, y = load_iris(return_X_y=True)
-            clf = LogisticRegression(random_state=0).fit(X, y)
-            print(clf.predict(X[:2, :]))
-            print(clf.predict_proba(X[:2, :]))
-            print(clf.score(X, y))
+            from scipy.linalg import lapack
+            lapack.dlamch('Epsilon-Machine')
             """
         )
